refactor: 코드리뷰 반영
- 로그인 및 회원가입 server action에서 비동기 함수로 수정
- GET 불필요한 application/json 삭제
- 드롭다운 use client 선언
- 드롭다운 프롭 styles에서 className으로 변경
- 헤더 z-index 적용
