+++ app/(auth)/login/_components/login-form.tsx
@@ -3,14 +3,13 @@
 import Button from "@/components/button/button";
 import { BasicInput } from "@/components/input-field/basic-input";
 import PasswordInput from "@/components/input-field/password-input";
+import { login } from "@/lib/apis/auth";
 import { loginSchema } from "@/schemas/auth";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { setCookie } from "cookies-next";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { SubmitHandler, useForm } from "react-hook-form";
-
-import { login } from "../../action";
 
 export interface LoginInputValue {
   email: string;
