+++ components/dropdown/dropdown.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import useClickOutside from "@/hooks/use-click-outside";
 import { ReactNode, createContext, useContext, useState } from "react";
 
@@ -69,7 +71,7 @@ Dropdown.Item = Item;
 
 interface DropdownButtonProps {
   children: ReactNode;
-  styles?: string;
+  className?: string;
 }
 
 /**
@@ -82,13 +84,13 @@ interface DropdownButtonProps {
  * @returns : 버튼 컴포넌트를 반환합니다.
  * @example : <Dropdown.Button>▽</Dropdown.Button>
  **/
-function Button({ children, styles }: DropdownButtonProps) {
+function Button({ children, className }: DropdownButtonProps) {
   const { handleDropdown, selected } = useDropdown();
 
   return (
     <button
       onClick={handleDropdown}
-      className={`${styles} flex w-full items-center`}
+      className={`${className} flex w-full items-center`}
     >
       {selected}
       {children}
@@ -99,21 +101,21 @@ function Button({ children, styles }: DropdownButtonProps) {
 // NOTE - Body
 interface BodyProps {
   children: ReactNode;
-  styles?: string;
+  className?: string;
 }
 
 /**
  * @author 김서영
  * 클릭 시에 나타나는 드롭다운 리스트를 감싸는 부분입니다
  * @param children : Item들이 children에 해당됩니다.
- * @param styles : 너비 및 배경색 등 추가적으로 적용될 스타일을 지정해주는 프롭입니다.
- * @example  <Dropdown.Body styles="w-36 bg-blue-200">...</Dropdown.Body>
+ * @param className : 너비 및 배경색 등 추가적으로 적용될 스타일을 지정해주는 프롭입니다.
+ * @example  <Dropdown.Body className="w-36 bg-blue-200">...</Dropdown.Body>
  **/
-function Body({ children, styles }: BodyProps) {
+function Body({ children, className }: BodyProps) {
   const { isDropdownOpen } = useDropdown();
 
   return isDropdownOpen ? (
-    <ul className={`${styles} absolute z-50`}>{children}</ul>
+    <ul className={`${className} absolute z-50`}>{children}</ul>
   ) : null;
 }
 
@@ -127,7 +129,7 @@ interface ItemProps {
  * @author 김서영
  * 드롭다운 선택 항목에 대한 컴포넌트입니다.
  * @param children : li 안에 포함될 내용을 적습니다
- * @param styles : 너비 및 배경색 등 추가적으로 적용될 스타일을 지정해주는 프롭입니다.
+ * @param className : 너비 및 배경색 등 추가적으로 적용될 스타일을 지정해주는 프롭입니다.
  * @param display : 선택 항목에 표시할 내용입니다. 없는 경우 children을 사용합니다.
  * @example  <Dropdown.Item><div className="flex gap-2"><p>Seo</p><span>Young</span></div></Dropdown.Item>
  **/
