+++ components/header/header.tsx
@@ -8,7 +8,7 @@ export default function Header() {
   const isLoggedIn = hasCookie("accessToken", { cookies });
 
   return (
-    <header className="fixed left-0 right-0 top-0 h-[60px] w-full bg-background-secondary px-4">
+    <header className="fixed left-0 right-0 top-0 z-50 h-[60px] w-full bg-background-secondary px-4">
       <div className="flex h-full items-center justify-between xl:mx-auto xl:max-w-[1200px]">
         {isLoggedIn ? <LoggedInHeader /> : <NoneLoginHeader />}
       </div>
